backend/requirements.txt
# ── Core Framework ────────────────────────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0
pydantic-settings>=2.2.0

# ── Database ──────────────────────────────────────────────────────────────────
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
supabase>=2.4.0

# ── Authentication ────────────────────────────────────────────────────────────
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9

# ── LLM Clients ───────────────────────────────────────────────────────────────
httpx>=0.27.0
openai>=1.23.0          # optional: only if LLM_PROVIDER=openai
anthropic>=0.25.0       # optional: only if LLM_PROVIDER=anthropic

# ── Embeddings & Vector Store ─────────────────────────────────────────────────
sentence-transformers>=2.7.0
faiss-cpu>=1.8.0        # use faiss-gpu for GPU support
numpy>=1.26.0

# ── ML Models ────────────────────────────────────────────────────────────────
scikit-learn>=1.4.0
xgboost>=2.0.3
lightgbm>=4.3.0
pandas>=2.2.0
prophet>=1.1.5          # Meta Prophet for time-series forecasting

# ── Document Parsing ──────────────────────────────────────────────────────────
pymupdf>=1.24.0         # PyMuPDF — fast PDF text extraction (preferred)
pypdf>=4.1.0            # Fallback PDF parser
python-docx>=1.1.0      # DOCX support

# ── Scheduling ────────────────────────────────────────────────────────────────
apscheduler>=3.10.4     # Cron-based report generation and model retraining

# ── Testing ───────────────────────────────────────────────────────────────────
pytest>=8.1.0
pytest-asyncio>=0.23.0
httpx>=0.27.0           # For FastAPI test client
